requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
tensorflow~=2.4.1
numpy~=1.20.2
pandas~=1.2.3
scikit-learn~=0.24.1
scipy~=1.6.2